add create index test
